requirements.txt
absl-py==0.13.0
astunparse==1.6.3
awscli==1.20.44
botocore==1.21.44
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.3
cycler==0.10.0
docutils==0.15.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
imutils==0.5.4
jmespath==0.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mss==6.1.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
tqdm==4.43.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
